requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
apify-client>=1.0.0
requests>=2.31.0
slack-sdk>=3.27.0
pyyaml>=6.0.0
jinja2>=3.1.0
python-dateutil>=2.8.0
